refacto(String): make String utils rather than plugin
