--- src/plugin/utils/tests/main.cpp
@@ -1,7 +1,0 @@
-#include <gtest/gtest.h>
-
-int main(int argc, char **argv)
-{
-    testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
